+++ test/support/my_app_cldr.ex
@@ -1,6 +1,0 @@
-defmodule MyApp.Cldr do
-  use Cldr,
-    locales: ["en", "fr"],
-    default_locale: "en",
-    providers: []
-end